telegram-bot/requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.11
uvloop==0.21.0; platform_system != 'Windows'

